requirements.txt
loguru==0.7.2
aiofiles==24.1.0
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.12.0
pylint==3.2.5
black==24.4.2
pytest==8.2.2
pytest-asyncio==0.23.7